frontend: carry granularity tiers into focus mode
Dialing a card to MED or HI and then expanding it snapped back to
mail-only: focus mode rendered raw lane events with no notion of
granularity. The three-tier merge (mail / summary history / raw
flow), the LO-MED-HI dial, and the summary and flow row renderers
now live in a shared timeline module used by both surfaces, so the
dial follows the agent into focus mode and back. Full-timeline focus
also gains the per-message focus glyph it was missing.
