+++ web/frontend/src/components/command/AgentCanvasCard.tsx
@@ -5,12 +5,13 @@
 
 import { useMemo, useRef, useState } from 'react';
 import { clsx } from 'clsx';
-import type { CommandEvent, CommandLane, FlowEvent } from '@/api/command.js';
+import type { CommandEvent, CommandLane } from '@/api/command.js';
 import type { AgentSummary } from '@/types/api.js';
 import { useCanvasStore } from '@/stores/canvas.js';
-import type { CardPosition, CardSize, Granularity } from '@/stores/canvas.js';
-import { useAgentFlow } from '@/hooks/useCommandFeed.js';
-import { useSummaryHistory } from '@/hooks/useSummaries.js';
+import type { CardPosition, CardSize } from '@/stores/canvas.js';
+import {
+  FlowRow, GranularityDial, SummaryRow, useAgentTimeline,
+} from './timeline.js';
 import { EventCard } from './EventCard.js';
 import { SteerComposer } from './SteerComposer.js';
 import type { ReplyTarget } from './SteerComposer.js';
@@ -20,79 +21,6 @@ import { agentTint, timeAgo } from './kinds.js';
 
 // Number of timeline items shown before the "older" fold.
 const VISIBLE_EVENTS = 8;
-
-// One entry in the merged three-tier timeline.
-type TimelineItem =
-  | { tier: 'mail'; ts: string; event: CommandEvent }
-  | { tier: 'summary'; ts: string; text: string; delta: string }
-  | { tier: 'flow'; ts: string; flow: FlowEvent };
-
-// Glyphs for flow event kinds, kept to quiet mono marks.
-const flowGlyph: Record<FlowEvent['kind'], string> = {
-  tool: '$',
-  thinking: '~',
-  text: '¶',
-  prompt: '›',
-};
-
-// FlowRow renders one high-granularity transcript event as a faint
-// mono line, expandable when a detail is present.
-function FlowRow({ flow }: { flow: FlowEvent }) {
-  const [open, setOpen] = useState(false);
-  return (
-    <button
-      type="button"
-      onClick={() => flow.detail && setOpen((v) => !v)}
-      className="flex w-full items-baseline gap-2 px-3 py-[3px] text-left hover:bg-[var(--c-paper)]"
-    >
-      <span className="w-3 shrink-0 text-center font-mono text-[10px] text-[var(--c-ghost)]">
-        {flowGlyph[flow.kind]}
-      </span>
-      <span className="min-w-0 flex-1">
-        <span
-          className={clsx(
-            'block truncate font-mono text-[10.5px] leading-4',
-            flow.kind === 'tool' ? 'text-[var(--c-mut)]' : 'text-[var(--c-faint)]',
-            flow.kind === 'thinking' && 'italic',
-          )}
-        >
-          {flow.kind === 'tool' && flow.detail
-            ? `${flow.label} ${flow.detail}`
-            : flow.label}
-        </span>
-        {open && flow.detail && flow.kind !== 'tool' && (
-          <span className="block whitespace-pre-wrap font-mono text-[10.5px] leading-4 text-[var(--c-mut)]">
-            {flow.detail}
-          </span>
-        )}
-      </span>
-      <span className="shrink-0 font-mono text-[9.5px] text-[var(--c-ghost)]">
-        {timeAgo(flow.timestamp)}
-      </span>
-    </button>
-  );
-}
-
-// SummaryRow renders one Haiku digest history entry: the medium tier.
-function SummaryRow({ text, delta, ts }: {
-  text: string; delta: string; ts: string;
-}) {
-  return (
-    <div className="flex items-baseline gap-2 px-3 py-1.5">
-      <span className="w-3 shrink-0 text-center font-mono text-[10px] font-bold text-[var(--c-green)]">
-        Δ
-      </span>
-      <span className="min-w-0 flex-1">
-        <span className="block text-[12px] leading-snug text-[var(--c-text2)]">
-          {delta || text}
-        </span>
-      </span>
-      <span className="shrink-0 font-mono text-[9.5px] text-[var(--c-dim)]">
-        {timeAgo(ts)}
-      </span>
-    </div>
-  );
-}
 
 export interface AgentCanvasCardProps {
   lane: CommandLane;
@@ -144,15 +72,6 @@ export function AgentCanvasCard({
   const setFocusedCard = useCanvasStore((s) => s.setFocusedCard);
   const setFocusedMessage = useCanvasStore((s) => s.setFocusedMessage);
   const setOpenDoc = useCanvasStore((s) => s.setOpenDoc);
-  const setGranularity = useCanvasStore((s) => s.setGranularity);
-
-  // Medium tier: Haiku summary history; high tier: transcript flow.
-  const { data: history } = useSummaryHistory(
-    agent.id, 20, granularity !== 'lo',
-  );
-  const { data: flowData } = useAgentFlow(
-    agent.id, granularity === 'hi',
-  );
 
   // Drag bookkeeping lives in a ref; only the store position renders.
   const dragRef = useRef<{
@@ -175,29 +94,7 @@ export function AgentCanvasCard({
 
   // Merge the three tiers into one timeline, newest first. Mail is
   // always present; summary history joins at MED; raw flow at HI.
-  const timeline = useMemo<TimelineItem[]>(() => {
-    const items: TimelineItem[] = lane.events.map((event) => ({
-      tier: 'mail', ts: event.created_at, event,
-    }));
-
-    if (granularity !== 'lo') {
-      for (const h of history ?? []) {
-        items.push({
-          tier: 'summary', ts: h.created_at,
-          text: h.summary, delta: h.delta,
-        });
-      }
-    }
-
-    if (granularity === 'hi') {
-      for (const f of flowData?.events ?? []) {
-        items.push({ tier: 'flow', ts: f.timestamp, flow: f });
-      }
-    }
-
-    items.sort((a, b) => (a.ts < b.ts ? 1 : -1));
-    return items;
-  }, [lane.events, history, flowData, granularity]);
+  const timeline = useAgentTimeline(lane, granularity);
 
   const visibleItems = useMemo(
     () => (showAll ? timeline : timeline.slice(0, VISIBLE_EVENTS)),
@@ -460,29 +357,8 @@ export function AgentCanvasCard({
         <span className="font-mono text-[9px] font-semibold uppercase tracking-[0.12em] text-[var(--c-dim)]">
           timeline
         </span>
-        <span className="ml-auto flex overflow-hidden rounded-md border border-[var(--c-hair)]">
-          {(['lo', 'med', 'hi'] as Granularity[]).map((g) => (
-            <button
-              key={g}
-              type="button"
-              title={
-                g === 'lo'
-                  ? 'Mail only'
-                  : g === 'med'
-                    ? 'Mail + activity summaries'
-                    : 'Everything incl. raw agent flow'
-              }
-              onClick={() => setGranularity(agent.id, g)}
-              className={clsx(
-                'px-1.5 py-px font-mono text-[9px] font-semibold uppercase',
-                granularity === g
-                  ? 'bg-[var(--c-ink)] text-white'
-                  : 'text-[var(--c-faint)] hover:bg-[var(--c-fill)]',
-              )}
-            >
-              {g}
-            </button>
-          ))}
+        <span className="ml-auto">
+          <GranularityDial agentId={agent.id} />
         </span>
       </div>
       <div className="scrollbar-thin min-h-0 flex-1 divide-y divide-[var(--c-fill)] overflow-y-auto">
